instant updates to search bar and organization view
